Refactor currency components with CurrencyHolder interfaces

+++ src/main/java/generics/Currency.java
@@ -2,11 +2,11 @@ package generics;
 
 public class Currency {
 
-    private int currencyDolarExchangeRate;
-    private int totalCurrencyDolarValue;
+    private double currencyDolarExchangeRate;
+    private double totalCurrencyDolarValue;
     private int currencyQuantity;
 
-    public Currency(int currencyDolarExchangeRate, int totalCurrencyDolarValue, int currencyQuantity) {
+    public Currency(double currencyDolarExchangeRate, double totalCurrencyDolarValue, int currencyQuantity) {
         this.currencyDolarExchangeRate = currencyDolarExchangeRate;
         this.totalCurrencyDolarValue = totalCurrencyDolarValue;
         this.currencyQuantity = currencyQuantity;
