+++ src/main/java/generics/standard/StandardCurrency.java
@@ -4,8 +4,7 @@ import generics.Currency;
 
 public class StandardCurrency extends Currency {
 
-    public StandardCurrency(int currencyDolarExchangeRate, int totalCurrencyDolarValue, int currencyQuantity) {
+    public StandardCurrency(double currencyDolarExchangeRate, double totalCurrencyDolarValue, int currencyQuantity) {
         super(currencyDolarExchangeRate, totalCurrencyDolarValue, currencyQuantity);
     }
-
 }
